requirements.txt
asgiref==3.6.0
Django==4.1.4
python-decouple==3.7
sqlparse==0.4.3
